+++ docs/Whats-New.md
@@ -501,6 +501,7 @@ New:
 - Forbidding parallel AI queues for specific TechnoTypes (by Starkku)
 - Nonprovocative Warheads (by Starkku)
 - Option to restore `PowerSurplus` setting for AI (by Starkku)
+- `FireOnce` infantry sequence reset toggle (by Starkku)
 - Penetration damage on garrisonable structures (by FS-21)
 - Map action 507 for printing a message with the remaining map objects (by FS-21)
 - ScriptType actions `10018` & `12003` for aircrafts attacks (by FS-21)